Broke down handRecognition into different files
