+++ hansroslinger/src/components/KonvaOverlay.tsx
