+++ hansroslinger/src/components/visuals/VegaLiteChartVisual.tsx
